Moved everything into gd.c
